+++ CHANGELOG.md
@@ -6,6 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+- Rename test-case-result keys:
+  * `:form` -> `:expected`
+  * `:result` -> `:actual`
+- Catch ExpectationFailed in `expect`, rethrow with additional data `:expected-message`, instead of passing `msg` into each assert-expr like in clojure.test.
+- Catch other Throwables in `expect`, wrap in ExpectationFailed as `:caught` data.
+- Require a docstring expression from `describe`, `it` and `expect-it`.
+- Add `*color*` dynamic var (set to lazytest.colorize system env, default to true) and make `colorize?` rely on it.
+- Move all relevant test case information into result object.
+- Move all relevant suite information into result object.
+- Filter suites/test cases when running a single var.
+- Add support for [nubank/matcher-combinators](https://github.com/nubank/matcher-combinators).
+- Sort test vars by line and column before running.
+- Rewrite runner to use type-based multimethod with report hooks.
+- Rewrite reporters to use new report hooks, allow for multiple hooks to be combined:
+  * `focused` prints if there's any focused tests.
+  * `summary` prints "Ran x test cases, N failures, Y errors".
+  * `results` prints failed and errored test cases, expected values, etc.
+  * `dots` prints `.` for passing test case, `F` for failure, and `E` for ERROR. Namespaces wrap test cases in parentheses: `(..F.)` Includes `focused`, `results`, and `summary`.
+  * `nested` prints each suite and test case on a new line, and indents each suite. Includes `focused`, `results`, and `summary`.
+  * `clojure-test` attempts to mimic clojure.test's basic output.
+- Add tests for reporters.
+
 ## 0.1.0 - 2024-06-09
 
 Updated original code to use deps.edn, tools.build, and other modern tooling.